Parent directories now work

+++ source/FAT32Directory.c
@@ -3,63 +3,7 @@
 #include <string.h>
 #include "../include/FAT32Directory.h"
 
-struct FAT32_entry_name_t
-{
-	char name[8];
-	char ext[3];
-};
-
-static struct FAT32_entry_name_t path_to_name(const char* path)
-{
-	struct FAT32_entry_name_t result;
-
-	// Extract the name
-	for (size_t i = 0; i < 8; ++i)
-	{
-		if (*path == '.' || *path == 0)
-		{
-			result.name[i] = ' ';
-		}
-		else
-		{
-			result.name[i] = *path;
-			path += 1;
-		}
-	}
-
-	// If we stopped adding the name because we reached a dot
-	if (*path == '.')
-	{
-		path += 1;
-	}
-
-	// Extract the extension
-	for (size_t i = 0; i < 3; ++i)
-	{
-		if (*path == 0)
-		{
-			result.ext[i] = ' ';
-		}
-		else
-		{
-			result.ext[i] = *path;
-			path += 1;
-		}
-	}
-
-	return result;
-}
-
-static struct FAT32_entry_name_t get_entry_name(const struct FAT32_directory_entry_t* entry)
-{
-	struct FAT32_entry_name_t result;
-	memcpy(result.name, entry->name, 8);
-	memcpy(result.ext, entry->ext, 3);
-
-	return result;
-}
-
-static FAT32_cluster_address_t get_entry_address(const struct FAT32_directory_entry_t* entry)
+FAT32_cluster_address_t FAT32_dir_get_entry_address(const struct FAT32_directory_entry_t* entry)
 {
 	FAT32_cluster_address_t result;
 	result.index_high = entry->first_cluster_index_high;
@@ -68,57 +12,76 @@ static FAT32_cluster_address_t get_entry_address(const struct FAT32_directory_en
 	return result;
 }
 
-static void set_entry_name(struct FAT32_directory_entry_t* entry, const struct FAT32_entry_name_t* name)
+void FAT32_dir_set_entry_address(struct FAT32_directory_entry_t* entry, FAT32_cluster_address_t address)
 {
-	memcpy(entry->name, name->name, 8);
-	memcpy(entry->ext, name->ext, 3);
+	entry->first_cluster_index_high = address.index_high;
+	entry->first_cluster_index_low = address.index_low;
 }
 
-static int compare_names(const struct FAT32_entry_name_t* lhs, const struct FAT32_entry_name_t* rhs)
+void FAT32_dir_get_entry_name(const struct FAT32_directory_entry_t* entry, char* outName)
 {
-	return strncmp(lhs->name, rhs->name, 8) || strncmp(lhs->ext, rhs->ext, 3);
-}
+	memset(outName, 0, FAT32_DIR_NAME_LEN);
 
-static void delete_entry(struct FAT32_directory_entry_t* entry)
-{
-	// If the entry is a subdirectory
-	if (entry->attribs & FAT32_DIR_ENTRY_ATTRIB_SUBDIRECTORY)
+	// Extract the name
+	for (size_t i = 0; i < 8; ++i, ++outName)
 	{
-		// Open 'er up
-		struct FAT32_file_t* file = FAT32_dir_open_entry(entry);
-
-		// Delete all entries in the subdirectory
-		struct FAT32_directory_entry_t subEntry;
-		while (FAT32_fread(&subEntry, sizeof(subEntry), 1, file))
+		if (entry->name[i] == ' ')
 		{
-			// If the subentry is a real entry
-			if (subEntry.name[0] != 0)
-			{
-				delete_entry(&subEntry);
-			}
+			break;
 		}
 
-		FAT32_fclose(file);
+		*outName = entry->name[i];
 	}
 
-	// Free the cluster chain
-	FAT32_free_cluster(get_entry_address(entry));
+	// Extract the extension
+	if (entry->ext[0] != ' ')
+	{
+		*outName = '.';
+	}
+
+	for (size_t i = 0; i < 3; ++i, ++outName)
+	{
+		if (entry->ext[i] == ' ')
+		{
+			break;
+		}
+
+		*outName = entry->ext[i];
+	}
+}
+
+void FAT32_dir_set_entry_name(struct FAT32_directory_entry_t* entry, const char* name)
+{
+	memset(entry->name, ' ', 8);
+	memset(entry->ext, ' ', 3);
+
+	char* target = entry->name;
+	for (; *name != 0; ++name, ++target)
+	{
+		if (*name == '.')
+		{
+			target = entry->ext;
+		}
+		else
+		{
+			*target = *name;
+		}
+	}
 }
 
 int FAT32_dir_get_entry(struct FAT32_file_t* dir, const char* name, struct FAT32_directory_entry_t* outEntry)
 {
-	// Get the name of the object
-	const struct FAT32_entry_name_t pathName = path_to_name(name);
-
     // Rewind the file
     FAT32_rewind(dir);
 
     // Loop until the entry is found
     while (FAT32_fread(outEntry, sizeof(struct FAT32_directory_entry_t), 1, dir))
     {
-		const struct FAT32_entry_name_t entryName = get_entry_name(outEntry);
+		// Get the name of the entry
+		char entryName[FAT32_DIR_NAME_LEN];
+		FAT32_dir_get_entry_name(outEntry, entryName);
 
-        if (compare_names(&pathName, &entryName) == 0)
+        if (!strcmp(name, entryName))
         {
 			// Rewind to the start of the entry
 			FAT32_fseek(dir, -(long)sizeof(struct FAT32_directory_entry_t), FAT32_SEEK_CUR);
@@ -132,7 +95,7 @@ int FAT32_dir_get_entry(struct FAT32_file_t* dir, const char* name, struct FAT32
 struct FAT32_file_t* FAT32_dir_open_entry(struct FAT32_directory_entry_t* entry)
 {
     // Construct the address
-	FAT32_cluster_address_t address = get_entry_address(entry);
+	FAT32_cluster_address_t address = FAT32_dir_get_entry_address(entry);
 
 	// If the entry is a subdirectory, open it with max size (directories are unsized)
 	if (entry->attribs & FAT32_DIR_ENTRY_ATTRIB_SUBDIRECTORY)
@@ -159,9 +122,6 @@ void FAT32_dir_close_entry(struct FAT32_directory_entry_t* entry, struct FAT32_f
 
 int FAT32_dir_new_entry(struct FAT32_file_t* dir, const char* name, FAT32_dir_entry_attribs_t attribs, struct FAT32_directory_entry_t* outEntry)
 {
-	// Create a name object for the given name
-	const struct FAT32_entry_name_t nameObj = path_to_name(name);
-
 	// Seek to the beginning of the file
 	FAT32_fseek(dir, 0, FAT32_SEEK_SET);
 
@@ -177,15 +137,13 @@ int FAT32_dir_new_entry(struct FAT32_file_t* dir, const char* name, FAT32_dir_en
 		insertPos = FAT32_ftell(dir);
 	}
 
-	// Set file properties TODO
-	set_entry_name(outEntry, &nameObj);
+	// Set file properties
+	FAT32_dir_set_entry_name(outEntry, name);
 	outEntry->attribs = attribs;
 	outEntry->size = 0;
 
 	// Create a cluster chain for the file
-	FAT32_cluster_address_t cluster = FAT32_new_cluster();
-	outEntry->first_cluster_index_high = cluster.index_high;
-	outEntry->first_cluster_index_low = cluster.index_low;
+	FAT32_dir_set_entry_address(outEntry, FAT32_new_cluster());
 
 	// Rewind to where we'll insert the file
 	FAT32_fseek(dir, insertPos, FAT32_SEEK_SET);
@@ -199,6 +157,34 @@ int FAT32_dir_new_entry(struct FAT32_file_t* dir, const char* name, FAT32_dir_en
 	return 1;
 }
 
+static void delete_entry(struct FAT32_directory_entry_t* entry)
+{
+	// If the entry is a subdirectory
+	if (entry->attribs & FAT32_DIR_ENTRY_ATTRIB_SUBDIRECTORY)
+	{
+		// Open 'er up
+		struct FAT32_file_t* file = FAT32_dir_open_entry(entry);
+
+		// Delete all entries in the subdirectory
+		struct FAT32_directory_entry_t subEntry;
+		while (FAT32_fread(&subEntry, sizeof(subEntry), 1, file))
+		{
+			// If the subentry is not a real entry or a system entry
+			if (subEntry.name[0] == 0 || subEntry.attribs & FAT32_DIR_ENTRY_ATTRIB_SYSTEM)
+			{
+				continue;
+			}
+
+			delete_entry(&subEntry);
+		}
+
+		FAT32_fclose(file);
+	}
+
+	// Free the cluster chain
+	FAT32_free_cluster(FAT32_dir_get_entry_address(entry));
+}
+
 int FAT32_dir_remove_entry(struct FAT32_file_t* dir, const char* name)
 {
 	// Get the entry to be removed
@@ -210,6 +196,12 @@ int FAT32_dir_remove_entry(struct FAT32_file_t* dir, const char* name)
 		return 0;
 	}
 
+	// Make sure its not a system entry
+	if (entry.attribs & FAT32_DIR_ENTRY_ATTRIB_SYSTEM)
+	{
+		return 0;
+	}
+
 	// Delete it
 	delete_entry(&entry);
 
@@ -217,4 +209,14 @@ int FAT32_dir_remove_entry(struct FAT32_file_t* dir, const char* name)
 	memset(&entry, 0, sizeof(entry));
 	FAT32_fwrite(&entry, sizeof(entry), 1, dir);
 	return 1;
+}
+
+void FAT32_dir_clear_entry(struct FAT32_directory_entry_t* entry)
+{
+	// Delete the entry (not as bad as it sounds)
+	delete_entry(entry);
+	entry->size = 0;
+
+	// Reallocate the cluster chain
+	FAT32_dir_set_entry_address(entry, FAT32_new_cluster());
 }
